+++ components/i2cmaster/src/i2cmaster.h
@@ -48,7 +48,7 @@ bool i2c_master_init(i2c_port_t i2c_num,
 int8_t i2c_wr(uint8_t dev_addr,
               uint8_t reg_addr,
               uint8_t *reg_data,
-              uint8_t cnt);
+              uint16_t cnt);
 
 /** <!-- i2c_rd {{{1 -->
  * @brief I2C master read function
@@ -61,7 +61,7 @@ int8_t i2c_wr(uint8_t dev_addr,
 int8_t i2c_rd(uint8_t dev_addr,
               uint8_t reg_addr,
               uint8_t *reg_data,
-              uint8_t cnt);
+              uint16_t cnt);
 
 // end of file {{{1
 // vim:ft=cpp:et:nowrap:fdm=marker
